refactor: move wire-protocol event types to shared layer
Channel shim and monitor imported TandemEvent, parseTandemEvent, and
formatEventContent/Meta from src/server/events/types.ts, crossing the
server boundary and pulling server code into self-contained tsup bundles
transitively. Move all wire-protocol types and functions to
src/shared/events/types.ts (with corrected relative import paths for
ReplyAuthor and generateEventId), fix two _exhaustive unused-local
warnings with void, and replace the server file with a barrel re-export
so server-internal consumers are unaffected.
